Added alert window to warn user for delete actions. Refactored error handling.

+++ frontend/src/pages/BrewResults.jsx
@@ -133,24 +133,41 @@ function BrewResults({ backendURL }) {
 
     // Handle user clicking Delete button on a record in the table
     const handleSingleDelete = async (result_id_to_delete) => {
-        const rid = parseInt(result_id_to_delete);
-        const deleteRes = await fetch(`${backendURL}/api/brewresults/${rid}`, { method: 'DELETE' });
-        if(deleteRes.status === 200) {
-            setBrewResults(brewResults => brewResults.filter(br => br.result_id !== rid));
-        } else {
-            console.log("Failed to delete Brew Result.");
+        if(window.confirm("Are you sure you want to delete this brew result?")) {
+            try {
+                const rid = parseInt(result_id_to_delete);
+                const deleteRes = await fetch(`${backendURL}/api/brewresults/${rid}`, { method: 'DELETE' });
+                if(deleteRes.status === 200) {
+                    alert("Brew result deleted successfully.")
+                    setBrewResults(brewResults => brewResults.filter(br => br.result_id !== rid));
+                } else {
+                    alert("Failed to delete brew result.");
+                    console.log("Failed to delete with status:", deleteRes.status);
+                }
+            } catch (err) {
+                console.error("Connection error:", err);
+                alert("Could not connect to the server.");
+            }
         }
     };
 
     // Handle the user completing the Bulk Delete form
     const handleBulkDelete = async (delete_recipe_id, delete_coffee_id) => {
-        const rid = parseInt(delete_recipe_id);
-        const cid = parseInt(delete_coffee_id);
-        const bulkDeleteRes = await fetch(`${backendURL}/api/brewresults/${rid}/${cid}`, { method: 'DELETE'});
-        if(bulkDeleteRes.status === 204) {
-            setBrewResults(brewResults => brewResults.filter(br => !((br.recipe_id === rid) && (br.coffee_id === cid))));
-        } else {
-            console.log("Failed to delete multiple Brew Result records");
+        if(window.confirm("Are you sure you want to delete all records associated with selected coffee and recipe ID?")) {
+            try {
+                const rid = parseInt(delete_recipe_id);
+                const cid = parseInt(delete_coffee_id);
+                const bulkDeleteRes = await fetch(`${backendURL}/api/brewresults/${rid}/${cid}`, { method: 'DELETE'});
+                if(bulkDeleteRes.status === 204) {
+                    setBrewResults(brewResults => brewResults.filter(br => !((br.recipe_id === rid) && (br.coffee_id === cid))));
+                } else {
+                    alert("Failed to delete multiple brew results");
+                    console.log("Failed to delete with status:", bulkDeleteRes.status);
+                }
+            } catch (err) {
+                console.error("Connection error:", err);
+                alert("Could not connect to the server.");
+            }
         }
     }
